test(init): migrate and consolidate remaining bot tests
